publish_audio_file_metadata_to_file_upload_completed_topic/requirements.txt
# Function dependencies, for example:
# package>=version
google-api-core>=2.2.2
